📌(frontend) pin styled-components & jest-styled-components versions
`jest-styled-components` published a new major version (7.0.0) which
contains different bugs, especially with TypeScript, and is not ready
for production. Wait until patch releases fix those issues to update.

Versions `6.x.x` of `jest-styled-components` do not work with the new
`5.0.0` version of `styled-components`, so we'll have to hold off on
this update too.