✨(front) implement use_subtitle_as_transcript behavior in video player
In the video player we should use subtitles instead of transcript if the
video property use_subtitle_as_transcript is set to true. To test this
feature we changed how timed text tracks are mocked in the video player
and choose to use ts-mock-imports like we did in other components.