✨(front) allow to change video quality while using ABR
Our user would like to choose the video quality and not let ABR decide
for them which one to use. But the main issue we had was the incompatibility
between plyr quality selector and dashjs. We found that plyr allow to
completely override the selector behaviour. The way we override it let
us change dashjs settings by setting the max bitrate the player can
choose, we don't rely anymore on plyr behaviour.