✨(playlists) allow defining precise playlist portability
Up to now, we could only allow either:
- any playlists of a consumer site to be portable to another consumer site
- one playlist of a consumer site to be portable to any other consumer site.

In some cases, we want to allow a playlist to be portable only to a
precise consumer site and only when we allow it. This can now be done by
creating a new playlist attached to the targeted consumer site, and defining
the new playlist as reachable from the playlist that we want to share.