✨(back) register xapi statement in a logger
We want to log xapi statement in a logger and also to configure a logger
by consumer site. We use for that a special logger name convention,
prefixing the domain name byt `xapi.`. For example for the consumer
site with domaine name foo.education you have to create a logger name
`xapi.foo.education` and configure as you want the associated handler
and formatter depending where you want to log the statements.