♻️(S3) use boto3 to generate upload form
To upload a file on s3, a form is crafted with a policy, signature and
more fields defining upload conditions. Since now all of this was
computed by us but we found that boto3 is able to do this job, it is
less error prone and we have less code to maintain.