✨(back) introduce live mode in video model
In order to manage live streaming we have to add new properties to video
model. The new properties are live_state representing the medialive
channel state and live_info containing all needed info to interact with
AWS once a medialive and mediapackage stack are created.
A first endpoint is added to initiate a complete live streaming stack on
AWS.