✨(AWS) create a lambda layer containing mediainfo binary
We want to use mediainfo at least in the encode lambda. To add a binary
in a lambda you zip it with the source code or create a layer and attach
this layer to a lambda. We choose to use a layer, it avoids to increase
the zip size and to upload it at every deployment.