♻️(AWS) revamp update-state code
The update state function was duplicated across all lambda using it.
When we wanted to update the code we had to copy/paste it
across all lambda. We created a dedicated module containing the
update-state code and we added it using package.json.
The signature is also changed. We sign the entire payload and the
signature is added in the X-Marsha-Signature request header.
Update-state now accepts an extraParameters object to send arbitrary
data to marsha.