♻️(back) refactor update-state endpoint
The update-state endpoint needs to be refactored. First the signature is
now in the X-marsha-Signature header and a new paramater extraParameters
is added.
This new field will contain arbitrary data specific to a model. The
generic way we use to update the state can no longer be used. We decided
to create a UploadableFileMixin containing fields and methods related to
upload management. The method update_upload_state can be overriden in
every model in order to manage data specific to the related model.