⬆️(frontend) update @testing-library/react to 10.4.3
While refactoring our testing-library powered tests to remove empty
wait (or waitFor) calls, we encountered a few ones which could not
be removed as that would break the related tests.

The issue was caused by lack of support for fake jest timers in
testing-library async utilities. It was fixed in dom testing library
v7.17.0 very recently, and incorporated in @testing-library/react
versions above v10.4.0.