🐛(front) fix how permissions are used to display dashboard
Permission object in JWT token was misused. To determine if a user can
see the instructor view it must have can_access_dashboard set to true
and to see the dashboard button in the instructor view it must have
can_update set to true. We did the opposite so when an instructor saw a
video from an other playlist it didn't see the message telling him that
the video is coming from an other playlist but he sees the student view.