📌(frontend) pin babel-plugin-react-intl <8
`format-js` introduced a new way to extract messages from source
files, using a CLI with purpose-made npx scripts. As a result, they
removed message-extracting capabilities from the babel plugin with
versions >8. It now only manages actual code transforms.

To perform this update, we need to install the CLI tools and rework
our l10n pipeline to use them instead of relying on a build
side-effect. This will be better done through a dedicated PR later
on.