✨(aws) save needed data to interact with cloudwatch
We plan to use the API to retrieve events in cloudwatch. We must save
some informations needed to do that. The first one is the logGroupName,
instead of adding another django settings we choose to pick it from an
event context and then send it to the backend in the update state
endpoint. We also need to save the medialive channel Arn to query the
events for a specific channel.