🐛(front) follow active transcript sentence without changing viewport
To synchronize active sentence of a transcript we were using
scrollIntoView function. It works but it also move the entire page to
have the sentence of the middle of the entire view port (all the
browser) and this a bug because the video can be not visible.
To solve this issue, we remove the usage of this function and instead we
compute where the scroll position should be and never focus on it.