🎨(front) create factories to improve tests readability and reliability
maintaining or creating new tests are not easy when video or document
types evolved. Tp resolve this we created a video and document factory.
These factories will be responsible to create video or document objects
respecting their respective types. Tests will use real objects and not
partial one with the risk to not really test with accurate data.