✨(auth/api) add drf token auth to retrieve video detail
This commit aim to add the drf token auth system. In this way,
we can easily retrieve (GET) video informations from a staff
user.

This commit was done because we sometimes need to retrieve data
from videos to enhance the ElasticSearch index with new data.
Without it, we would have to manually forge LTI tokens to get
video data.