🐛(back) fix sentry configuration to have env and release info
When an error is reported to sentry fron the backend application, the
environment and release info are not set correctly. The bug is explained
in issue #723. To fix it, we don't call anymore method in the post_setup
method because this method is a classmethod and can't access to method.