✨(aws) create a lambda responsible to route event to a medialive lambda
In the previous medialive-event, when a medialive event was fired, all
medialive lambdas were triggered on all environments leading to unwanted
behavior. The solution here is to create only one lambda for all
environments responsible to route the event to the good medialive event.
In this lambda we must fetch the channel payload to use its tags so the
channel is included in the lambda call and the medialive lambda doesn't
need anymore to use the aws sdk.