♻️(aws) use input to execute lambda_migrate instead of env var
To know the migrations to run, an environment variable called MIGRATIONS
containing a comma separated list and this string is split in the lambda
function. There is an easier way to run this lambda using variable
complex type. The migrations terraform variable is now a string list and
this variable is used as an input when the lambda is invoked. No more
environment variable, no more string parsing, less code to maintain.