✨(back) store timed text track origin extension
We wanto to add in the timed text track payload the link to the original
file. This original file can be downloaded and we need to know its
extension to add it to the downloaded file name. A new property
extension is added to the timed text track model and the extension is
saved from the update state endpoint.