♻️(aws) fix regex changing timed text track destination key
The regex changing the timed text track destination key was replaced by
a simple text replacement. This modification is wrong because we also
remove other part of the source key. This issue was not detected by the
tests because they were not relevant at all, this PR refactor them with real
case values.