📌(front) temporary ping mux.js to version 5.7.0
We have identified an issue in videojs. The bug is in a nested
dependency, mux.js. A new version of mux.js containing a fix is
available and we don't to wait a new videojs release to have this fix. A
temporary solution is to add a resolution in our package.json file. This
resolution will be removed once a new release of videojs available.