👷(circle) build and publish lambda image
We want to automate the build and the publish of the lambda image. For
this we added 2 new jobs. One responsible to build the image on every
push and an other one to publish the image on every merge on master
branch and when a new tag is added to the repo.