🐛(aws) use absolute path to register presets
Lambda run in the docker image are not launched from the lambda targeted
root but from the LAMBDA_TASK_ROOT directory. In the lambda-configure
function, all the presets are registered using the relative path from
the root of the lambda, this is not correct anymore. We have to use the
absolute path.