⚡️(medialive) switch to QVBR rate control mode
Up to now we were using CBR (Constant BitRate) control mode. It is the
easiest one to implement. We decided to switch to QVBR, it's harder to
implement but we can save a lot of bandwidth and also space without
losing quality.
To understand QVBR and which value to use we follow this guide :
https://docs.aws.amazon.com/mediaconvert/latest/ug/cbr-vbr-qvbr.html