🔥(aws) remove terraform state management
We choose to remove the state management in marsha.
First it was not configurable.
Secondly making it configurable is not easy. It is forbidden to use
variable in a backend block.
By removing it, every one is responsible to manage the state and can
choose how to manage it, This is the first aws lock vendor removal.