✨(aws) create a lambda to manage mediapackage harvest events
When a live is ended, an harvest job will be created, Once this one
succeeded we want to trigger a lambda running ffmpeg to transcode the
HLS video into an MP4 container and save this newly created video in the
destination bucket.
In a lambda it is not possible to write more then 512MB of data in /tmp.
To resolve this issue, we attached an EFS (Elastic File System). To
attach this EFS we also have to configure a VPC (Virtual Private Cloud).
The EFS is connected to the VPC and both are configured in the lambda.