🏗(frontend) rework frontend rendering to make room for a new app
Marsha as it exists has one frontend, the LTI frontend it uses to
manage and make available videos and documents in the context of LTI
requests.

We want to start allowing user actions in a more general context,
outside of the lifecycle of an LTI request. The first use-case for
this is mass uploads of videos. To that end, we're starting up a new
marsha site, which will take the form of a new frontend root.

We can create a space for this new site by adding a view, rendered
into the "/" root path, as LTI requests are behind a dedicated path.
We then load the same frontend index the LTI view uses and render
a different app ("LTI" or "site") depending on AppData contents.

We chose to do it this way to maximize code reuse between both
apps, as a lot of the logic and tools ought to be the same, as
well as the styling.