✨(frontend) create a base layout for the marsha site
Unlike the marsha LTI app, which is the existing marsha frontend
as a whole, the new site is not meant to be viewed inside an LTI
iframe. As a result, it must occupy the whole frame and needs a
basic page structure to do so.

Use grommet builtins to lay it out on the page and make it the
base from which we'll build our actual pages.