✨(api) add an endpoint to get the current user
To build up the interface for the new marsha site frontend, we need
to know who the current user is, and what organizations they belong
to, as those are the entry points to video uploads.

We added a user serializer and an organization access serializer
with the organization name stapled on to avoid making additional
requests.