✨(frontend) add a data source for the current user
As we build the marsha site, we're using an actual django user
instead of an LTI session and a bunch of JWT claims. As a result,
we need to use our new whoami endpoint to get some information
about the current user.

Create a User type matching what our API user serializer returns
and build out a small utility to access and cache it with zustand.