✨(api) open a list API route for playlists
Part of the role of the new marsha site, apart from the flagship
feature of mass uploading videos, is to enable users to create
playlists and manage them, so the videos can be made available
through LMSes.

Before we can add or manage playlists, we need to be able to
list them, for a given user (everything they have access to), and
for a given organization.