♻️(api) refactor video ViewSet to add new methods
We want to add new methods to the video ViewSet that will be
needed in the new marsha site.
Before we do that, we want to change the ViewSet to inherit from
ModelViewSet so we can use DRF builtins.

We're using a get_permissions method to keep the behavior the
same for now and only open the actions as we need them.