♻️(aws) split mediapackage lambda in sub function
to speed the mediapackage lambda function execution and prevent to reach
a timeout we decided to split the function in sub function invoked via
the same lamda. The index is responsible to call the good sub functions
:
- harvest: first function responsible to compute how many chunks to
create and invoke the transmux function.
- transmux: transmux a chunk of the video and check if all the chunks
are generated. If yes the concat function is invoked.
- concat: concat all chunk to create the final video and also generates
the thumbnail linked to the given resolution. Check if all video are
concatenated and invoke the upload function if it's the case
- upload: upload all generated videos and thumbnails to s3