✨(api) open a delete endpoint for videos on the API
With the Marsha LTI application, there was no need to enable users
to delete videos at all. The way LTI works, the LTI request creates
the individual video, and we considered it not the scope of the
individual video content management system to handle deletions.

Now that we want to enable mass upload of videos through the new
marsha site, we're actually building a CRUD management interface.
There, it makes sense to let users "clean up" unused or unwanted
videos as part of their management UI.