🔥(front) remove usage of dash
We removed the usage of dash. We have to remove it from the urls in the
video schema and its usage in all video players. We can also remove the
dash player used when plyr player is used.
By default HLS is used and it's the player responsability to use it or
fallback to MP4 sources.