♻️(frontend) rework <Loader /> to expand use cases
The existing <Loader /> in marsha was only appropriate for use
in the LTI view as it forced a `fixed` position with an overlay
covering the whole viewport. It also lacked some accessibility
related elements.

We made available a new <Spinner /> component from the same file,
which enables users to use only the animated spinner without the
whole full-screen overlay.

We also made some accessibility improvements to said <Spinner />
which we refactored into the basis for <Loader />.