♻️(api) update existing list requests with DRF pagination
Before we added the default DRF pagination to our settings, there
already existed one api route for a list of resources. We need to
update its tests and call sites in the frontend to make sure it
can work from the pagination-supporting object DRF returns, rather
than a plain list.