✨(front) manage live to vod publication
Publishing a live in VOD must be a manual validation. For this we
created a dedicated component <DashboardVideoHarvested /> displaying
watch and publish button and managing CTA on publish button. When the
publish button is clicked, an update on the API is made and the store is
refreshed with the new state.