✨(front) create a videojs plugin to change mp4 quality played
When HLS is not supported we fallback on MP4 files. Videojs reads the
first source item found and it's not possible to manually switch the
video quality.
We created a plugin to handle this case. The plugin creates a menu list
containing all resolution available. allowing to start playing with a
default quality. When the user chooses a different quality, the source
is changed and we seek in the video to the previous registered time.