♻️(frontend) rework <ErrorComponent /> to reuse it in site
The general use error component already in marsha is meant to
take up the whole screen in the context of the LTI iframe Marsha
operates in.

This is not appropriate for the new marsha site, where we want
most errors to be colocated with the content they are linked with
instead.
We set out to reuse the existing component instead of creating
a separate generic error component.