🔧(front) enable videojs useDevicePixelRatio VHS option
By default, Videojs Http Streaming component disable an option called
useDevicePixelRatio. This option take the device pixel ratio into
account when doing rendition switching. This means that if you have a
player with the width of 540px in a high density display with a device
pixel ratio of 2, a rendition of 1080p will be allowed. This option
enabled combined with the undocumented option limitRenditionByPlayerDimensions
set to true by default will allow to use a higher rendition on devices
allowing it.