✨(api) open an endpoint to retrieve a playlist
As part of our new Marsha site, we want to be able to retrieve
a single playlist so we can display its name and all related
meta-data on the individual playlist view.

For now, only administrators are granted access to these views,
so we're likewise restricting the API route to administrators.