➕(frontend) add react-query to our dependencies
We are building a bunch of CRUD views with lists of items that
can be heavy, and a lot of detailed views in the marsha site.

As we get away from lighter UI with more defined constraints, and
the marsha site evolves in a more open-ended way in the future, it
will make more sense to have a general-purpose cache layer like
what react-query provides, instead of a custom-build and hand-managed
store.