🐛(front) remove HLS source in <VideoPlayer />
When we set the HLS source in the <video /> tag, the browser try to load
it but fails if the browser is not able to do it. Videojs will load it
because it has polyfills to play HLS source in a browser havinf
MediaSource extension. The solution is to remove the HLS source in the
<video /> tag and configure directly videojs player with accurate
sources dependening if media sources are available or not.