✨(frontend) add a set of components to manage breadcrumbs
We're using react-router-dom to build the new Marsha site. We're
creating a page tree, using nested switches to enable easy routing
and to keep our "place" in the app even if we reuse the same routes
in different places in the site.

To help users navigate this and provide them affordances for going
back and forth, we decided to use breadcrumbs. This is a somewhat
automatic implementation that attempts to make it as easy as possible
to add the crumbs dynamically throughout the app.