✨(static) add a frontend component & a static file for icons
As we build the new marsha site, we needed a way to display SVG
icons throughout the app.

In addition to creating an icons.svg file in statics with the first
two icons we're using, we created an <Icon /> component to handle
using the correct static location & making sure the icon is accessible.