✨(frontend) create an <UploadManager /> component
We want to use a common element to handle uploads in both the
existing LTI application and the new Marsha site. To that end,
we're creating an <UploadManager /> to hold the uploads state
and take care of uploads.

This is intended to replace both the `upload` site effect and
the object progress state. It should provide a more complete
state that allows consumers to know everything they need about
ongoing and complete uploads.