♻️(back) avoid concurrent requests on video live update state endpoint
A medialive channel has many pipelines. One by RTMP input. Each pipeline
triggers events and these events are listen by the medialive lambda. In
a first time we thought that this events are always triggers twice (one
by pipeline) and choose to listen only pipeline 0. In reality, this
assert is not always true. Sometimes, I didn't found why, only one
pipelin is triggering an event so we can miss them if they are not on
the pipeline 0. To solve this issue, all the events are listened again
and we update first the live state in the update_live_state endpoint. If
the state is already updated, we stop the process and return a status
200 to stop lambda retries mechanism.