♻️(frontend) rework <UploadStatusPicker /> to use upload manager
The existing version of the <UploadStatusPicker /> component worked
off `upload_state` as a prop. This was a simplification as the LTI
frontend used an `UPLOADING` state that does not exist on the backend,
and optimistically set `PROCESSING` on objects where it had not received
such a state update for the server.

As we build out the Marsha site, and strive to find as much common
ground between the site and the LTI app as possible, we needed the upload
states to exactly match what exists in the real data (especially as the
site is constantly refreshing its cache from the server).

In turn, the upload manager comes in to make sure we used the most
up-to-date information for states where the server is not involved
(during an upload), and when the lambda => server => client loop has
not happened yet (processing state before an object refresh from the API).

As we were working on this, we realized the `<UploadStatus />` sub-component
was not serving much of a purpose anymore and just got rid of it.