♻️(frontend) update untitled object titles with file name
There is currently a "fix" in the <UploadForm /> that updates
objects which can receive a title with the file name as the object
title upon download completion.

This provides titles for objects without relying on user input. Is
is not the way we intend to proceed with the new Marsha site. In our
opinion, the LTI app will probably evolve in the same direction
later on.

However, for now we need to provide feature parity as we simply
rework the LTI app to use <UploadManager />. To that end, we added
a component at root tasked with applying this mutation on all
relevant objects when uploads successfully end.