♻️(back) change status in check_live_state command
In the check_live_state management command, when a live need to be
stopped, we set the live state to STOPPED. That was good before we made
the live to VOD feature. Since we have to first set with STOPPING state.
The STOPPED will by set by calling the update live state endpoint.