✨(back) install and configure whitenoise
In order to speed up deployment we don't want to store and serve assets
from S3. We choose to use whitenoise, a django lib allowing to compress
and cache assets and serve them using a CDN. We already have a
cloudfront distribution for this purpose, we just need to modify it to
use this new configuration.