♻️(frontend) update <UploadForm /> & <UploadField />
As we rework marsha's upload related components to use the
new <UploadManager />, the <UploadForm /> & <UploadField />
component are part of this effort.

We mostly lightened the form by using the manager, as we do not
intend to reuse it on the marsha site. The upload field itself
on the other hand is intended to be reused in the site and thus
needs to remain more feature-light while connecting with the
manager.