♻️(frontend) update dashboard elements for upload manager
As we moved to the <UploadManager /> to handle uploads in marsha's
LTI app, we removed the "fictional" UPLOADING upload state on
uploadable objects.

As a result, we had some refactoring to do in the dashboard elements
that made use of it to make sure the UI remained unchanged and was
still functional.