✅(back) update state signature tests
All the tests about an update state using a signature process used an
already computed signature. On a specific test, we randomly generate the
request body, it's not possible to process the signature. We choose to
create a dedicated function generating a heash given a secret and a
message and use this function in all the tests to generate the
signature.