🗃(migrations) make lti_id optional on Video and Document
As we first built marsha, the lti_id field was essential to the
way we let users interact with videos and documents through LTI
requests.

Marsha evolved, and we still store the LTI id for videos and
documents created through LTI, but it is no longer a necessary
element to have a functioning video/document. It also creates an
issue as objects created through the new marsha site cannot by
definition have an LTI id upon creation.

We can just make the field optional on the model without breaking
anything. Objects created through LTI will do without it while objects
created through LTI will still have a value there in case we have a
use for it.