♻️(frontend) make plyr styles LTI app specific
The Plyr player sizing is expressed in `vw` or viewport width
points. This is appropriate in the LTI app which lives in a small
iframe.

It produces broken video sizes in the marsha site as `vw` is not
the appropriate unit to use there.

We can add wrapper to apply a class around the LTI view so we can
target only the LTI app with style rules.