✨(frontend) update <Icon /> component with size param
As part of the new uplods view, we want to displat larger icons
in list items for uploads. To allow this more easily, we're updating
the <Icon /> component to allow custom icon sizes, different from
the default `1rem` size we're using for all icons that are used
close to text.