✨(frontend) add react-query based hooks for API interactions
As we build out the marsha site, we want to interact with the API
without slowing down our app nor building out a complicated cache
layer.

To achieve this, we can build out a set of helpers around react-query
that will enable us to call the API with no boilerplate from any
of our components.

This provides us with an automatic stale-while-revalidate cache
layer so our views load quickly even when they are waiting for
requests to complete.