✨(frontend) create an upload view & its related indicator
As uploading a bunch of videos is the first use case of the new
marsha site, we want to help users know what is currently going
on with their uploads.

We can leverage our <UploadManager /> to add a live indicator to
the sidebar, and a more comprehensive view that lists all ongoing
and completed file uploads.