✨(frontend) add a component to display lists of videos
Displaying a list of videos is a need that is recurring in the new
marsha site and that will probably arise in other places.

As a result, we want to build a flexible component to consistently
show that list, with differing parameters depending on the use case.
We intend to upgrade it with search, filtering and pagination later
on as we continue to build out Marsha.

We also needed a video badge to compute and display a user-facing
status for a given video.