✨(frontend) add a 2-step form to create a new video
In the Marsha LTI app, videos are automatically created when the
LTI request is made to marsha by the LMS. This is not the case in
the Marsha site, where we need to manually create the videos from
the frontend.

To help users find their videos later on, we need to make it
mandatory to add a title for the video (and the playlist, if the
context does not make it obvious). We made the form 2-step so they
can first input the relevant values before they pick a file to
upload. The third step is just a helpful message with a button to
open the form for another video.