♻️(back) create function to manage common app data in views
All views share commin app data. Since now, these app data are craft
when we need to return them and repeated several times. This lead to
case where app data can be missing, we add one in one case and not in
the others and can lead to a bug. To prevent this we refactor how
app_data are computed.