🙈(frontend) update `data` gitignore entry to cover only root folder
`data` is in the gitignore entries as it is used to manage statics
with the docker image. However, we also use it as a folder name in
our frontend, in which we want to be able to add new files.

Limit the entry to only gitignore data in the root folder so it
does not cover the frontend folder.