📝(terraform) add a bit of information on `getopt` for mac users
Script related to the AWS part of Marsha's stack are broken for
Mac users by default, as the version of `getopt` available by
default in macOS is not compatible with versions on other Unixes.

It can be fixed by installing the GNU version and adding it to the
PATH. We decided to add a small disclaimer in the docs so new users
don't waste time trying to figure this out.