♻️(api) clean up permissions for api endpoints
In an earlier change, we introduced the `ObjectPkMixin` to have
an easy way for our custom permissions to access the current call's
relevant object id, without resorting to a match on the string in the
permission itself.

We're updating the permissions that we're still using a string match
to use the view mixin instead. We also took this opportunity to remove
needless try/catch blocks around queryset `.exists()` calls.