♻️(api) rework permission implementations to enable combinations
Permissions as implemented by Django-rest-framework are broken, as
we see it. The semantics of the OR operator for permissions are not
correct, or at least do not match what the developer would expect,
and result in undefined behavior when combining permissions that use
both `has_permission` and `has_object_permission`.

Our proposed solution is to never use `has_object_permission`, so
combination semantics are clearer and permissions can be combined at
will.

We settled on always using `has_permission` rather than
`has_object_permission` as some methods do not use `has_object_permission`
and would have forced us to make exceptions. Also, it should be faster
(eg. less database requests) for most API calls.