✨(front) display chat box without video player
As an instructor I would like to see only the chat box during a live
without seeing the video itself. A route is added redirecting to the
chat component and a button is added in the dashboard. An instructor can
see the chat only or the video and the chat in the same page.