🐛(front) initialize only once conversejs instance
Everytime the Chat component was mounted, conversejs was also intialized
and can lead to non restorable session error. To resolve this issue,
converse instance must be initialized once and then this instance must
be inserted in the dom everytime the component is mounted. Converse
provide a helper `converse.insertInto` allowing to mount converse in an
existing html element.