🚨(front) remove warning fetch not found
In createVideoJsPlayer tests we render a real Video component. In the
video component we fetch available timed text track languages in the
useEffet hook. As we are in a node context, fetch is not defined and a
warning was display. To resolve this issue we choose to mock the call to
getChoices.