♻️(front) use video object from state in Chat component
Since now the xmpp info injected in the chat component where retrieved
in the video object and it works fine like this if the component is
mounted in the VideoPlayer component. But if we mounted directly using
the chat only feature, the data used come from the appData and can be
outdated. We must retrieve the video from the state to be sure to have
up to date data.