🐛(front) reset upload manager state before starting a new upload
It is impossible to chain multiple upload for the same object. Once an
upload made you can not replace the same object if you mistaken for
example. Once on the UploadForm component again you are redirected to
the dashboard. To prevent this issue we choose to reset the state if an
entry already exists with the same object id.