♻️(back) create a video property to store live type
A new property live_type is added in the video model. This property is
responsible to store the kind of live is created. Adding a new property
instead od storing it in the live_info object allow us to update this
data easily from the API. Doing this the front can change the type once
crated and switch between types available.