✨(front) display a waiting message while live is not available
When a live begins, you have to wait a few seconds before the hls
manifest is created and available. If the player starts fetching this
missing manifest, it fails and never restart. We choose to poll the
manifest while it returns a 200 status code and then the player is
created. A waiting message is display during this period.