✨(front) create a component to manage a public video dashboard
The VideoPlayer was the main component managing the video player but
also every feature linked to the video public part like the download
links, displaying transcripts and the chat. We decided to create a
component dedicated to manage a public video dashboard and depending the
video type it displays the needed components. Splitting it will allow us
to choose how to display the components independently of each other.