✨(front) manage video live states in <PublicVideoDashboard />
The front application is now responsible to manage video live state. A
video with live_state not null has is_ready_to_show value set to true
but the manifest is not always available. While the live is not
available, a waiting message is displayed and the player appears once
the live avaiable.