✨(front) add zustand store to manage participant
In order to manage participant to join a discussion we need new store to
keep thei state and share it accross multiple components. The
useJoinParticipant store is used in the dashboard by an instructor to
accept, reject or kick a partipant in a discussion. The
useParticipantWorkflow is used by the participant itself wanted to join
a discussion.