✨(front) converse plugin managing join the discussion
The join discussion feature is manage through the xmpp server. Messages
are sent to instructors and students to join or not the discussion. For
this we created a new converse plugin and this plugin exposes new
function in the converse object otself allowinf to use them where we
need to.