✨(front) sync participant state between instructors
When a participant want to join a discussion, all instructors receive
the message and all can accept, reject or kick the participant. But the
action is not sync between instructors so if one accept the participant,
nothing will happen on others dashboard. This commit sync the state
broadcasting messages between instructors.