✨(site) update urls for the marsha site
The marsha site exists on the root path of its domain, but
currently redirects itself to /site on the frontend side.

We can replace the simple `path` calls on the Django side with
a `re_path` that catches all requests and let the frontend resolve
the requested URL. We also do not need to add "/site" in front of
our frontend routes.