🩹(back) public token has no LTI context
LTI context is a notion related to LTI. If a resource is visualized
from a public context, without an LTI Passeport, we shouldn't
specify any `context_id`. This notion is confusing and not
properly true. To avoid any misunderstanding, we erase from
the token the key `context_id` in a public context.