🐛(front) force xapi initialized action for a live video
For a live video, no event to detect when the video is fully initialized is
triggered before the first "play" action. To mitigate this, we can call
"initialize" on the first "interact" action and we don't log this interaction.
The first interact action is when the first quality to play is chosen, the
default one. To choose it, all quality available must be read in the HLS
manifest. So we can consider at this time that the video is initialized.