🐛(front) listen video play event instead of playing
The video html element dispatches 2 differents events related to
playback. `play` and `playing`. In a VOD context using one or the other
lead to the same result. Moreover, when we used plyr player, we follow
its documentation[1] to implement video events and listen the correct one
in order to send XAPI request, and this documentation is not really the
same as the MDN one[2].
In a live context, the `playing` is triggered a new video fragment is
loaded. So we can have thousands of `played` XAPI events and this is
wrong. We must listen video `play` event. And this event is also the
good one to listen in a VOD context.

[1]: https://github.com/sampotts/plyr#events
[2]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/video#events