♻️(back) remove LTIUser helper and directly use JWT token
The class LTIUser in the lti module was created to ease the use of the
JWT token. It was a bad idea. First, the name of this class is not
relevant at all, a JWT token can be used outside a LTI context and it
can introduce a lot of confusion. Moreover, using directly the JWT token
is good enough. The class LTIUser is removed and the JWT token is
directly used to build an xapi statement.