✨(JWT) add email in the token
In the introduction of a feature that will enable scheduled videos,
we want to notify users by email when they register. Email is not
necessarily present in the LTI request, that's why it wasn't added
so far in the JWT token. It is still a piece of information that can
be known and it will be accurate to use this specific email to link
an LTI user.