🔧(aws) use amazon lambda image from ECR public repo
The amazon/aws-lambda-nodejs host on docker hub use the ARM64
architecture, it seems it is a bug. docker hub is a mirror of ECR public
repo host on amazon. We choose to directly use the image hosted on ECR
to avoid this kind of issue. The image can be found at this url
https://gallery.ecr.aws/lambda/nodejs