♻️(front) change chunk filename pattern to a more simple one
With new version of webpack we can use the contenthash in the chunk
filenane. The contenthash is the hash of the file content, it ensures
that the filename change everytime its content changes. This filename
improve cache allowing to use the same file between two releases if the
content is exactly the same.