⚡️(back) create a custom staticfiles storage class
The staticfiles storage we use with whitenoise create a manifest using
the django manifest mechanism and then compress all post processed
files. In our case, for 15700 files are post processed for 3383 static
files ! For one js file generated by webpack, we have at the end 6
versions of it. Our new staticfiles storage ignore all js and map files
generated by webpack, they are not post processed by the django
manisfest mechanism but we still compress to take the advantage of
whitenoise. Finally we have 4500 post processed file for 3383 static
files. It decreases the time to build the image.