✨(xapi) track download activity
When a user downloads a document or a video, a xapi statement is sent to
track this activity. We use a hack to combine the `<a href` and onclick
behaviour. If we send the xapi statement at the same time the download
is made, browsers stop the xapi fetch. The fetch is postponed to the
window blur event which happens once the download begins.