✨(front) redirect instructor to jitsi prejoin page once conference left
When an instructor leaves a jitsi conference we don't want to display
the jitsi homepage as we do for now. We destroy and recreate the jitsi
session and we force him to wait on the prejoin page. Here the
instructor can join again the conference or quit marsha by closing the
browser.