♻️(front) each resource is responsible to redirect the user on load
Depending on the loaded resource, we want to redirect the user to different
routes. Since now not having this architecture was enough because we
didn't have specific cases based on the resource itself. Now we want to
redirect the user on the dashboard if it's a video live and he has the
update permission.