♻️(back) create AWS stack if not existing when starting live
Instead of firstly create all the AWS stack and then start the medialive
channel once created we want to do it in the start live endpoint. If the
AWS stack is not created, we create it then we wait while the medialive
channel is IDLE and then we start it.