✨(back) pause live instead of permanently stopping it
The update-live-state endpoint now will pause the live when receiving a
`stopped` event. The AWS stack is not deleted nor harvested anymore on
this event. To avoid race condition, when a retry is made for example,
we record the aws request id and return earlier the response if the same
request id is played multiple times. This endpoint is faster now, we
don't call AWS endpoint nor wait for the stack to be deleted so we will
have less race conditions.