♻️(back) reopen chat to non members on restart
Before having the start/stop feature for a live, the chat was open when
the live saw started and closed when the live was stopped. Closing the
chat means restrict its access to members only. As there is no member on
the prosody server, nobody, except the admin, can access the chat. With
the start/stop feature, the chat must be reopened on the next start.