✨(front) detect when a live is resumed to update video store
When a live is paused, the manifest contains `#EXT-X-ENDLIST`. This tag
means no more media segments will be added to the manifest. Fortunately,
mediapackage will remove this tag once the live resumed, meaning the
live_state has changed in RUNNING and we can fetch the video object to
refresh its state in the application store. To do this, we fetch the
manifest to parse it while this tag is present. Once removed, we stop
fetching and the video is updated in the store.