✨(frontend) register for a scheduled webinar
Now that instructors can create a scheduled webinar, we want students
to be able to subscribe to these events. Before registering, users
can see the information of the future webinar. Depending on the data
contained in the user’s token, he will have to add his email to
subscribe or just click on the register button. Once the starting
date of the webinar is reached, the user will have access to the
waiting screen.