🐛(back) no validator for field starting_at in the model Video
Video model had a validator associated with the field starting_at.
This one was triggered as soon as starting_at field had a value. We
only want this field to be checked if its value gets changed. We
decided that this field could be changed outside the API context,
for instance from the administration panel. The control of this
field is now only realized in the serializer.